# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  dto.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ApiError'
      meta:
        $ref: '#/definitions/dto.ResponseMeta'
      success:
        type: boolean
    type: object
...
  dto.ChainAnalysis:
    properties:
      criticalPathPosition:
        type: string
      floatHours:
        type: integer
      longestChain:
        type: integer
    type: object
...
  dto.DependencyInfo:
    properties:
      dependencyId:
        type: string
      dependencyType:
        type: string
      dependsOnTaskId:
        type: string
      isBlocking:
        type: boolean
      lagHours:
        type: integer
      status:
        type: string
    type: object
  dto.DependencySection:
    properties:
      direct:
        items:
          $ref: '#/definitions/dto.DependencyInfo'
        type: array
      indirect:
        items:
          $ref: '#/definitions/dto.IndirectDependency'
        type: array
    type: object
...
  dto.DependentInfo:
    properties:
      dependencyType:
        type: string
      isBlocked:
        type: boolean
      taskId:
        type: string
    type: object
  dto.DependentsSection:
    properties:
      direct:
        items:
          $ref: '#/definitions/dto.DependentInfo'
        type: array
      indirect:
        items:
          $ref: '#/definitions/dto.IndirectDependency'
        type: array
    type: object
...
  dto.IndirectDependency:
    properties:
      depth:
        type: integer
      path:
        items:
          type: string
        type: array
    type: object
...
  dto.PriorityBreakdown:
    properties:
      businessValue:
        type: integer
      criticalPathWeight:
        type: integer
      deadlineUrgency:
        type: integer
      dependencyImpact:
        type: integer
      projectPriority:
        type: integer
    type: object
  dto.PriorityFactors:
    properties:
      blockedTasksCount:
        type: integer
      businessValue:
        type: integer
      daysUntilDue:
        type: integer
      deadlineUrgency:
        type: integer
      isOnCriticalPath:
        type: boolean
      projectPriority:
        type: integer
    type: object
...
  dto.ResponseMeta:
    properties:
      hasMore:
        type: boolean
      nextCursor:
        type: string
      page:
        type: integer
      perPage:
        type: integer
      requestId:
        type: string
      timestamp:
        type: string
      total:
        type: integer
    type: object
...
  dto.SubtaskProgress:
    properties:
      progressPercentage:
        type: integer
      status:
        type: string
      taskId:
        type: string
      title:
        type: string
    type: object
...
  dto.TaskDependenciesResponse:
    properties:
      chainAnalysis:
        $ref: '#/definitions/dto.ChainAnalysis'
      dependencies:
        $ref: '#/definitions/dto.DependencySection'
      dependents:
        $ref: '#/definitions/dto.DependentsSection'
      taskId:
        type: string
    type: object
  dto.TaskPriorityResponse:
    properties:
      breakdown:
        $ref: '#/definitions/dto.PriorityBreakdown'
      calculatedAt:
        type: string
      factors:
        $ref: '#/definitions/dto.PriorityFactors'
      rank:
        type: integer
      taskId:
        type: string
      universalPriorityScore:
        type: integer
    type: object
...
  dto.TaskProgressResponse:
    properties:
      actualHours:
        type: number
      completedAt:
        type: string
      dueDate:
        type: string
      estimatedHours:
        type: number
      progressPercentage:
        type: integer
      remainingHours:
        type: number
      startDate:
        type: string
      status:
        type: string
      subtaskProgress:
        items:
          $ref: '#/definitions/dto.SubtaskProgress'
        type: array
      taskId:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Xephyr Project Management Intelligence API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Xephyr API
  version: "1.0"
paths:
...
  /dependencies/tasks/{taskId}:
    get:
      consumes:
      - application/json
      description: Get all dependencies for a specific task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Include indirect dependencies
        in: query
        name: includeIndirect
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskDependenciesResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get task dependencies
      tags:
      - dependencies
...
  /priorities/tasks/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the priority score and breakdown for a specific task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskPriorityResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get single task priority
      tags:
      - priorities
...
  /progress/tasks/{taskId}:
    get:
      consumes:
      - application/json
      description: Get detailed progress for a task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskProgressResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get task progress
      tags:
      - progress
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"